Refactor(middleware): change file location
